Add `AutoconfigParser` interface
