+++ feature/autodiscovery/autoconfig/src/main/kotlin/app/k9mail/autodiscovery/autoconfig/AutoconfigParserException.kt
@@ -1,0 +1,3 @@
+package app.k9mail.autodiscovery.autoconfig
+
+class AutoconfigParserException(message: String, cause: Throwable? = null) : RuntimeException(message, cause)
